# RT_ComOps-openapi.yaml
openapi: "3.1.0"
info:
  title: "RT_ComOps API"
  description: "RT_ComOps API"
  version: "1.0.0"
servers:
  - url: "https://RT_ComOps"
paths:
...
  /employees/invite:
    post:
      summary: "POST employees/invite"
      operationId: "inviteEmployee"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrganizationMember"
...
components:
  schemas:
    Instant:
      type: "object"
      properties: { }
...
    InviteRequest:
      type: "object"
      properties:
        email:
          type: "string"
          nullable: true
        roleId:
          type: "string"
          format: "uuid"
          nullable: true
        permissionIds:
          type: "array"
          nullable: true
          items:
            type: "string"
            format: "uuid"
        agencyId:
          type: "string"
          format: "uuid"
          nullable: true
    OrganizationMember:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
          nullable: true
        organizationId:
          type: "string"
          format: "uuid"
          nullable: true
        userId:
          type: "string"
          format: "uuid"
          nullable: true
        userEmail:
          type: "string"
          nullable: true
        userFirstName:
          type: "string"
          nullable: true
        userLastName:
          type: "string"
          nullable: true
        agencyId:
          type: "string"
          format: "uuid"
          nullable: true
        agencyName:
          type: "string"
          nullable: true
        roleId:
          type: "string"
          format: "uuid"
          nullable: true
        roleName:
          type: "string"
          nullable: true
        isActive:
          type: "boolean"
          nullable: true
        joinedAt:
          nullable: true
          $ref: "#/components/schemas/Instant"
...